requirements.txt
numpy
librosa
scikit-learn
matplotlib
scipy

torchaudio
pyannote.audio
python-dotenv
flask
PyJWT